+++ utils/accounts.ts
@@ -1,7 +1,6 @@
 // IMPORTANT: we cannot import hardhat directly here
 // because it will cause a circular dependency
 import { Account } from "./types";
-import * as dotenv from "dotenv";
 import { Wallet } from "ethers";
 
 export const loadEnv = (name: string, defaultValue: string) => {
@@ -20,9 +19,9 @@ export const loadAddress = (name: string, privateKey: string) => {
       const wallet = new Wallet(privateKey);
       address = wallet.address;
       console.log(name + ": " + address);
-    } catch (error:any) {
+    } catch (error: any) {
       console.error(
-        `Error deriving address from private key for ${name}: ${error.message}`
+        `Error deriving address from private key for ${name}: ${error.message}`,
       );
     }
   }
@@ -37,30 +36,30 @@ export const ACCOUNTS: Account[] = [
     name: "QB",
     privateKey: loadPrivateKey(
       "QB",
-      "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"
+      "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80",
     ),
   },
   {
     name: "GRANTOR",
     privateKey: loadPrivateKey(
       "GRANTOR",
-      "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d"
+      "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d",
     ),
   },
   {
     name: "GRANTEE",
     privateKey: loadPrivateKey(
       "GRANTEE",
-      "0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a"
+      "0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a",
     ),
   },
-].map((account:Account) => {
+].map((account: Account) => {
   // Check if the address is not already present
   if (!account.address) {
     // Derive the address using the loadAddress function
     account.address = loadAddress(account.name, account.privateKey);
   }
-  return account
+  return account;
 });
 
 // map of account name -> account
@@ -69,7 +68,7 @@ export const NAMED_ACCOUNTS = ACCOUNTS.reduce<Record<string, Account>>(
     all[acc.name] = acc;
     return all;
   },
-  {}
+  {},
 );
 
 // map of account name -> account address
@@ -78,7 +77,7 @@ export const ACCOUNT_ADDRESSES = ACCOUNTS.reduce<Record<string, string>>(
     all[acc.name] = acc.address;
     return all;
   },
-  {}
+  {},
 );
 
 // flat list of private keys in order
